Change README.md to fit new repo objectives
As discussed via private chat over the last days, `ref` should
serve the purpose of containing a book list that Daque members
find interesting to
a) develop viare, or
b) become better developers

The changes in this commit reflect this new objective by modifying
the proccess of note taking. Books are now categorized as general or
viare-specific.